restructure the folders and formate code properly, also add launch screen
