+++ MintynDigitalBank/MintynDigitalBank/Delegates/SceneDelegate.swift
@@ -8,10 +8,10 @@
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    
     var window: UIWindow?
-
-
+    
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = scene as? UIWindowScene else {
             return
@@ -22,27 +22,27 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window?.rootViewController = initialViewController
         window?.makeKeyAndVisible()
     }
-
-    func sceneDidDisconnect(_ scene: UIScene) {
-   
-    }
-
-    func sceneDidBecomeActive(_ scene: UIScene) {
-     
-    }
-
-    func sceneWillResignActive(_ scene: UIScene) {
     
+    func sceneDidDisconnect(_ scene: UIScene) {
+        
     }
-
+    
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        
+    }
+    
+    func sceneWillResignActive(_ scene: UIScene) {
+        
+    }
+    
     func sceneWillEnterForeground(_ scene: UIScene) {
-      
+        
     }
-
+    
     func sceneDidEnterBackground(_ scene: UIScene) {
-       
+        
     }
-
-
+    
+    
 }
