+++ MintynDigitalBank/MintynDigitalBank/ViewControllers/HomePageViewController/AdvertCollectionViewCell.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 class AdvertCollectionViewCell: UICollectionViewCell {
- 
+    
     lazy var imageView: UIImageView = {
         let image = UIImageView()
         image.translatesAutoresizingMaskIntoConstraints = false
@@ -17,21 +17,21 @@ class AdvertCollectionViewCell: UICollectionViewCell {
         image.layer.cornerRadius = 4
         return image
     }()
-
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
         setupViews()
         self.layer.cornerRadius = 10
         self.layer.masksToBounds = true
     }
-
+    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
+    
     func setupViews() {
         [imageView].forEach { addSubview($0) }
-
+        
         NSLayoutConstraint.activate([
             imageView.topAnchor.constraint(equalTo: self.topAnchor),
             imageView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
