+++ MintynDigitalBank/MintynDigitalBank/ViewControllers/LegalVC/LegalVC+Extension.swift
@@ -1,51 +1,13 @@
 //
-//  LegalViewController.swift
+//  LegalVC+Extension.swift
 //  MintynDigitalBank
 //
-//  Created by Decagon on 17/06/2022.
+//  Created by Decagon on 20/06/2022.
 //
-
 import UIKit
+import Foundation
 
-class LegalViewController: UIViewController {
-    // MARK: - Properties (Created Using IIFE: Immediately Invoked Function Expression)
-    let reusableCell = "legalPageCell"
-    let legalData = ["Mintyn Account- Terms & Conditions", "Mintyn Privacy Policy", "Saving Goals- Terms & Conditions"]
-    //  Navigation Button
-    lazy var navigationBtn: UIButton = {
-        let navBtn = UIButton()
-        let config = UIImage.SymbolConfiguration(pointSize: 25, weight: UIImage.SymbolWeight.bold)
-        navBtn.translatesAutoresizingMaskIntoConstraints = false
-        navBtn.setImage(UIImage(systemName: "chevron.backward", withConfiguration: config), for: .normal)
-        navBtn.addTarget(self, action: #selector(dismissTheView), for: .touchUpInside)
-        return navBtn
-    }()
-    // Header Title
-    lazy var navTitleHeader: UILabel = {
-        let navTitle = UILabel()
-        navTitle.translatesAutoresizingMaskIntoConstraints = false
-        navTitle.text = "Legal"
-        navTitle.textColor = UIColor.CustomColor.textColorGray
-        navTitle.font = customFont(font: .robotoBlack, size: 24)
-        return navTitle
-    }()
-    lazy var legalListTableView: UITableView = {
-        let legalListTableView = UITableView()
-        legalListTableView.translatesAutoresizingMaskIntoConstraints = false
-        legalListTableView.delegate = self
-        legalListTableView.dataSource = self
-        legalListTableView.separatorStyle = .none
-        legalListTableView.register(UITableViewCell.self, forCellReuseIdentifier: reusableCell)
-        return legalListTableView
-    }()
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .systemBackground
-        setUPView()
-    }
-    @objc func dismissTheView() {
-        dismiss(animated: true)
-    }
+extension LegalViewController {
     func setUPView() {
         let uiViews = [navTitleHeader, navigationBtn, legalListTableView]
         for uiView in uiViews {
@@ -64,7 +26,7 @@ class LegalViewController: UIViewController {
             legalListTableView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -12),
             legalListTableView.topAnchor.constraint(equalTo: navTitleHeader.bottomAnchor, constant: 15),
             legalListTableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
-            ])
+        ])
     }
 }
