+++ MintynDigitalBank/MintynDigitalBank/ViewControllers/SettingsPageVC/SettingsPageVC+Extension.swift
@@ -1,39 +1,13 @@
 //
-//  SettingsPageViewController.swift
+//  SettingsPageVC+Extension.swift
 //  MintynDigitalBank
 //
-//  Created by Decagon on 17/06/2022.
+//  Created by Decagon on 20/06/2022.
 //
-
 import UIKit
+import Foundation
 
-class SettingsPageViewController: UIViewController {
-    // MARK: - Properties (Created Using IIFE: Immediately Invoked Function Expression)
-    let settingsData = SettingsTimeTableList.getListOfMeal()
-    let reusableCell = "settingsPageCell"
-    // Header Title
-    lazy var navTitleHeader: UILabel = {
-        let navTitle = UILabel()
-        navTitle.translatesAutoresizingMaskIntoConstraints = false
-        navTitle.text = "Settings"
-        navTitle.textColor = UIColor.CustomColor.textColorGray
-        navTitle.font = customFont(font: .robotoBlack, size: 24)
-        return navTitle
-    }()
-    lazy var settingListTableView: UITableView = {
-        let settingsTable = UITableView()
-        settingsTable.translatesAutoresizingMaskIntoConstraints = false
-        settingsTable.delegate = self
-        settingsTable.dataSource = self
-        settingsTable.separatorStyle = .none
-        settingsTable.register(UITableViewCell.self, forCellReuseIdentifier: reusableCell)
-        return settingsTable
-    }()
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .systemBackground
-        setUpViews()
-    }
+extension SettingsPageViewController {
     func setUpViews() {
         let items = [navTitleHeader, settingListTableView]
         for item in items {
