+++ MintynDigitalBank/MintynDigitalBank/ViewControllers/SettingsPageVC/SettingsPageViewController.swift
@@ -1,0 +1,38 @@
+//
+//  SettingsPageViewController.swift
+//  MintynDigitalBank
+//
+//  Created by Decagon on 17/06/2022.
+//
+
+import UIKit
+
+class SettingsPageViewController: UIViewController {
+    // MARK: - Properties (Created Using IIFE: Immediately Invoked Function Expression)
+    let settingsData = SettingsTimeTableList.getListOfMeal()
+    let reusableCell = "settingsPageCell"
+    // Header Title
+    lazy var navTitleHeader: UILabel = {
+        let navTitle = UILabel()
+        navTitle.translatesAutoresizingMaskIntoConstraints = false
+        navTitle.text = "Settings"
+        navTitle.textColor = UIColor.CustomColor.textColorGray
+        navTitle.font = customFont(font: .robotoBlack, size: 24)
+        return navTitle
+    }()
+    lazy var settingListTableView: UITableView = {
+        let settingsTable = UITableView()
+        settingsTable.translatesAutoresizingMaskIntoConstraints = false
+        settingsTable.delegate = self
+        settingsTable.dataSource = self
+        settingsTable.separatorStyle = .none
+        settingsTable.register(UITableViewCell.self, forCellReuseIdentifier: reusableCell)
+        return settingsTable
+    }()
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .systemBackground
+        setUpViews()
+    }
+}
+
